add idn-area api connection

+++ data/src/main/java/com/reader/bacalagi/data/base/BaseResponse.kt
@@ -2,7 +2,7 @@ package com.reader.bacalagi.data.base
 
 data class BaseResponse<T>(
     val status: String,
-    val statusCode: Long,
+    val statusCode: Int,
     val message: String,
     val data: T?
 )
